Adding connection retries for keystone adapter
Due to fact that the same state cinder are run on the
all control nodes and this state means restarting of
apache therefore raise condition occurs and as a consequence
BadStatus lines errors is raised whiting keystone connection. So
in order to avoid this connect_retries is set to make keystone client
reconnect to the server.

Related-PROD: PROD-19729

https://github.com/openstack/keystoneauth/blob/master/keystoneauth1/adapter.py#L45

Change-Id: I8ba2de1197d07b9c11994f3c7e89223430fcbc8f